move db.utils to deprecated namespace
