+++ db/queries/base.py
@@ -7,7 +7,7 @@ from db.columns.operations.select import get_column_name_from_attnum
 from db.tables.operations.select import reflect_table_from_oid
 from db.transforms.operations import apply
 from db.transforms import base
-from db.utils import execute_pg_query
+from db.deprecated.utils import execute_pg_query
 from db.metadata import get_empty_metadata
