+++ db/tables/operations/select.py
@@ -1,7 +1,7 @@
 from sqlalchemy import Table, select, join
 
 from db.connection import exec_msar_func
-from db.utils import execute_statement, get_pg_catalog_table
+from db.deprecated.utils import execute_statement, get_pg_catalog_table
 
 BASE = 'base'
 DEPTH = 'depth'
