+++ db/tests/types/test_array.py
@@ -1,5 +1,5 @@
 import pytest
-from db.utils import execute_pg_query
+from db.deprecated.utils import execute_pg_query
 from db.functions.base import ColumnName, Noop, ArrayContains
 from db.functions.packed import (
     ArrayLengthEquals, ArrayLengthGreaterThan, ArrayLengthGreaterOrEqual,
