+++ db/tests/types/test_json.py
@@ -1,5 +1,5 @@
 import pytest
-from db.utils import execute_pg_query
+from db.deprecated.utils import execute_pg_query
 from db.functions.base import ColumnName, Noop, JsonArrayContains
 from db.functions.packed import (
     JsonLengthEquals, JsonLengthGreaterThan, JsonLengthGreaterorEqual,
