+++ db/types/custom/email.py
@@ -3,7 +3,7 @@ from sqlalchemy.types import UserDefinedType
 
 from db.types.base import MathesarCustomType
 from db.types.custom.underlying_type import HasUnderlyingType
-from db.utils import ignore_duplicate_wrapper
+from db.deprecated.utils import ignore_duplicate_wrapper
 
 DB_TYPE = MathesarCustomType.EMAIL.id
